:sparkles: api: Move theoricalEndTime to domain model
