+++ api/src/certification/session-management/domain/usecases/get-session-for-supervising.js
@@ -1,7 +1,3 @@
-import dayjs from 'dayjs';
-
-import { DEFAULT_SESSION_DURATION_MINUTES } from '../../../shared/domain/constants.js';
-
 /**
  * @typedef {import('./index.js').SessionForSupervisingRepository} SessionForSupervisingRepository
  * @typedef {import('./index.js').CertificationBadgesService} CertificationBadgesService
@@ -13,30 +9,7 @@ import { DEFAULT_SESSION_DURATION_MINUTES } from '../../../shared/domain/constan
  * @param {CertificationBadgesService} params.certificationBadgesService
  */
 const getSessionForSupervising = async function ({ sessionId, sessionForSupervisingRepository }) {
-  const sessionForSupervising = await sessionForSupervisingRepository.get({ id: sessionId });
-
-  sessionForSupervising.certificationCandidates.forEach(_computeTheoricalEndDateTime);
-
-  return sessionForSupervising;
+  return sessionForSupervisingRepository.get({ id: sessionId });
 };
 
 export { getSessionForSupervising };
-
-/**
- * @param {CertificationCandidateForAd} certificationBadgesService
- */
-function _computeTheoricalEndDateTime(candidate) {
-  const startDateTime = dayjs(candidate.startDateTime || null);
-  if (!startDateTime.isValid()) {
-    return;
-  }
-
-  let theoricalEndDateTime = startDateTime.add(DEFAULT_SESSION_DURATION_MINUTES, 'minute');
-
-  if (candidate.isStillEligibleToComplementaryCertification) {
-    const extraMinutes = candidate.enrolledComplementaryCertification.certificationExtraTime ?? 0;
-    theoricalEndDateTime = theoricalEndDateTime.add(extraMinutes, 'minute');
-  }
-
-  candidate.theoricalEndDateTime = theoricalEndDateTime.toDate();
-}
